requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==4.2.27
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.0.0
idna==3.11
paystackapi==2.1.3
psycopg2-binary==2.9.11
dj-database-url==1.1.0
PyJWT==2.10.1
requests==2.32.5
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
reportlab==3.7.4
